Pin react-router-dom to specific version to make it work with npm
npm doesn't look at our `yarn.lock` which specifies we want `react-router-dom` version `6.0.0-beta.0`.  Instead it looks at `package.json` which says `^6.0.0-beta.0` which grabs beta version 8, and then something fails to compile at

https://github.com/earthstar-project/letterbox/blame/main/src/Dashboard.tsx#L70

...saying it can't find `path` on that match object.

Temporary solution for now is to pin to the specific version we want so it works with `npm install`.  I tested that this works.